backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
aiofiles==23.2.1
sqlalchemy==2.0.23
python-dotenv==1.0.0